+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Remove unused avro property from `build.gradle.kts`
 * Trim schemaless data to only pass the fields that are in the table
 * Allow bypassing the schema validation
+* Detect if table schema has changed and refresh the schema
 
 # 1.2.4
 * Adjusting underlying client version to 0.7.0